Join multiple networks around a target node ID

This function joins multiple networks and resticts the joined network to interactions involving a target node ID of interest

joinNetworks(id, network, intA.col = "SymbolA", intB.col = "SymbolB")

Arguments

id

character. A gene or protein ID around which a joined network should be constructed.

network

list. Each element of the list is a data.frame representing a network. That means each row of a data.frame represents an interaction.

intA.col

character. Name of the column that stores the ID of the first interactor in each data.frame of argument network.

intB.col

character. Name of the column that stores the ID of the second interactor in each data.frame of argument network.

Value

A data.frame containing the joined network.

Examples

 library(BioPlex)
 bp.293t <- BioPlex::getBioPlex(cell.line = "293T", version = "3.0")
#> Using cached version from 2023-01-14 23:15:28
 bp.hct <- BioPlex::getBioPlex(cell.line = "HCT116", version = "1.0")
#> Using cached version from 2023-01-14 23:49:23
 netl <- list(HEK = bp.293t, HCT = bp.hct)

 jnet <- joinNetworks(id = "CDH2", network = netl)